Can't actually use usize for these operations. Swap to u64.
Network operations like this need to send a specified number of bytes
(duh) or clients of different architectures will send incorrect numbers
of bytes to each other.

Idk if u32 or u64 is better. Just going with u64 because why not!

+++ src/headers.rs
@@ -46,7 +46,7 @@ impl ConnectionSend for Header {
 impl ConnectionGet for Header {
     fn get(con: &mut Connection) -> Result<Self, String> {
         let mut action = [0 as u8; 1];
-        let mut length = [0 as u8; mem::size_of::<usize>()];
+        let mut length = [0 as u8; mem::size_of::<u64>()];
 
         // TODO check that the right amount of bytes were read
         con.get(&mut action)?;
